separate font and icons from screen package
