+++ screen/basic_screen.go
@@ -212,43 +212,6 @@ func (screen *GameScreen) SetInputTextAndRender(text string) {
 	screen.Render()
 }
 
-func (screen *GameScreen) pylonIcon() string {
-	// return "🔶"
-	return "🔷"
-}
-
-func (screen *GameScreen) goldIcon() string {
-	return "💰"
-}
-
-func (screen *GameScreen) blueBoldFont() func(string) string {
-	return screen.colorFunc(fmt.Sprintf("%v+bh:%v", 117, 232))
-}
-
-func (screen *GameScreen) brownBoldFont() func(string) string {
-	return screen.colorFunc(fmt.Sprintf("%v+bh:%v", 181, 232))
-}
-
-func (screen *GameScreen) brownFont() func(string) string {
-	return screen.colorFunc(fmt.Sprintf("%v:%v", 181, 232))
-}
-
-func (screen *GameScreen) regularFont() func(string) string {
-	return screen.colorFunc(fmt.Sprintf("255:%v", bgcolor))
-}
-
-func (screen *GameScreen) greyFont() func(string) string {
-	return screen.colorFunc(fmt.Sprintf("%v:%v", 181, 232))
-}
-
-func (screen *GameScreen) blinkBlueBoldFont() func(string) string {
-	return screen.colorFunc(fmt.Sprintf("%v+B:%v", 117, bgcolor))
-}
-
-func (screen *GameScreen) inputActiveFont() func(string) string {
-	return screen.colorFunc(fmt.Sprintf("0+b:%v", bgcolor-1))
-}
-
 func (screen *GameScreen) redrawBorders() {
 	io.WriteString(os.Stdout, ansi.ColorCode(fmt.Sprintf("255:%v", bgcolor)))
 	screen.drawBox(1, 1, screen.Width()-1, screen.Height()-1)
